Refactor project testing routes into separate modules
